Доработки в DocumentController, внедрение Aspect конфигурация для Kafka (с использованием inbox/outbox паттернов)

+++ backend/src/main/java/ru/abdullaeva/sber/backend/documents/repository/DocumentRepository.java
@@ -4,8 +4,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.abdullaeva.sber.backend.documents.model.Document;
 
-
+/**
+ * Репозиторий для работы с документами
+ */
 @Repository
-public interface DocumentRepository extends JpaRepository<Document, Integer> {
+public interface DocumentRepository extends JpaRepository<Document, Long> {
 }
